+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3628,7 +3628,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KQXRQYBGHHQVRNGDZBDARG2B"),
+  id: z.string().default("01KQY7B4TEPM38V2DDVBYYMF1Q"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),