+++ packages/orq-rc/RELEASES.md
@@ -8869,3 +8869,13 @@ Based on:
 - [typescript v4.9.0-rc.25] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.25 - packages/orq-rc
+
+## 2026-05-06 08:47:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.26] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.26 - packages/orq-rc